Add shard handling code
A shard is an image file consisting of a tgz and a header that lists
dependencies. While the dependencies might tempt one to think of a shard as a
sort of "package," a shard is best thought of as an image that has been
differentially compressed against other images.

This commit introduces the Shard class which can be used to load and save shard
files, though presently one must construct the internal tgz manually.

Signed-off-by: Casey Dahlin <casey.dahlin@gmail.com>